requirements.txt
openai>=1.0.0
mss>=9.0.1
pyautogui>=0.9.54
Pillow>=10.0.0
mcp>=1.0.0
pyperclip>=1.8.0
